+++ messages/en/admin.php
@@ -291,6 +291,8 @@ return [
     'con_amend_editorial'     => 'Allow <strong>editorial hints</strong> <small>(a separate text field)</small>',
     'con_amend_globalalt'     => 'Allow <strong>global alternatives</strong> <small>(amendments that completely replace the motion)</small>',
     'con_proposal_procedure'  => 'Show the <strong>proposed procedure</strong> on a separate page',
+    'con_show_private_notes'  => 'Users can write <strong>Private Notes</strong>',
+    'con_open_links_new_tab'  => 'Open PDFs and external links in new browser-tabs/windows',
     'con_proposal_tt'         => 'Allows admins to propose a procedure for motions and amendments. Has to be explicitly activated per motion type.',
     'con_pp_inline'           => '<strong>Show changes inline</strong> by default if the proposed procedure proposes changes',
     'con_pp_inline_tt'        => 'Only works for motions, and only if they do not have amendments',