+++ tests/Support/AcceptanceTester.php
@@ -75,7 +75,7 @@ class AcceptanceTester extends Actor
     public function gotoMotion(bool $check = true, string $motionSlug = '321-o-zapft-is'): MotionPage
     {
         if (is_numeric($motionSlug)) {
-            /** @var \app\models\db\Motion $motion */
+            /** @var Motion $motion */
             $motion     = Motion::findOne($motionSlug);
             $motionSlug = $motion->getMotionSlug();
         }