+++ views/motion/_view_text.php
@@ -137,9 +137,9 @@ foreach ($sections as $i => $section) {
 if ($right === '') {
     $out = $main;
 } else {
-    $out = '<div class="row" style="margin-top: 2px;"><div class="col-md-8 motionMainCol">';
+    $out = '<div class="motionTwoCols"><div class="motionMainCol">';
     $out .= $main;
-    $out .= '</div><div class="col-md-4 motionRightCol">';
+    $out .= '</div><div class="motionRightCol">';
     $out .= $right;
     $out .= '</div></div>';
 }