+++ web/typescript/frontend/MotionShow.ts
@@ -7,6 +7,11 @@ class MotionParagraph {
     private readonly lineHeight: number;
 
     constructor(private $element: JQuery) {
+        const s = location.hash.split('#amendment');
+        if (s.length == 2) {
+            $(".bookmarks .amendment" + s[1]).first().scrollintoview({top_offset: -100});
+        }
+
         this.$paraFirstLine = $element.find(".lineNumber").first();
         this.lineHeight = this.$paraFirstLine.height();